OAK-6271 - Support for importing index files
Do not clean the output directory if it contains the import directory

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1802236 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-run/src/main/java/org/apache/jackrabbit/oak/index/IndexCommand.java
@@ -249,7 +249,12 @@ public class IndexCommand implements Command {
 
     private static void setupDirectories(IndexOptions indexOpts) throws IOException {
         if (indexOpts.getOutDir().exists()) {
-            FileUtils.cleanDirectory(indexOpts.getOutDir());
+            if (indexOpts.isImportIndex() &&
+                    FileUtils.directoryContains(indexOpts.getOutDir(), indexOpts.getIndexImportDir())) {
+                //Do not clean directory in this case
+            } else {
+                FileUtils.cleanDirectory(indexOpts.getOutDir());
+            }
         }
         cleanWorkDir(indexOpts.getWorkDir());
     }
